Make sure all AJAX calls specify JSON as a return value.
Add a DataONE theme router to disable Tools, Plans, Registry, etc.

If a tools route doesn't exist, do not add a link to the EML info from
the Download Contents table in the Metadata view.

Add a Download Contents table div element to the MetadataIndex template
so that the table is always inserted into the DOM.

Properly URL escape every ID in each call made by the metadata views.