Added another two helper functions to the SolrResult model to retrieve source IDs and derivation IDs of a SolrResult based on its index info.
Added a function to the Package model to retrieve all the provenance
information about all members of a data package. It finds the resource
map ID of all external sources and derivations and finds all other
objects in the index that have provenance fields pointing to a member of
the package.