Add geographic, temporal, and taxonomic coverage classes, and begin to flesh them out.
For geographic coverages, we want to support point, line, and polygon features, but this is difficult in EML (but not impossible with some rule bending).  Plan to use GeoJSON as the internal format, and use toXML() to provide the EML.  Use toGeoJSON() to translate EML into GeoJSON.

For temporal coverage, we'll initially punt on alternative time scales and use string types for begin and end dates.  If endDate is not present, we create a singleDateTime in the EML, otherwise we create a rangeOfDates.  We also won't support repeated singleDateTimes for now.