Override the synced attribute on the new EML doc
This makes it so the metadata editor shows up for new documents. The
synced property is set because the DataPackage add event checks the
synced attribute and only renders the metadata if true. Here, I've
overridden the property once but we may want to consider overriding the
property at init-time because we create new EML instances in other
places in the code. I'm not sure about this though.