Work on support for file upload:
Listen for edit button clicks and drag/drop events on metadata table rows.  Add the handleAddFiles(), addFiles(), showdropzone(), and hideDropzone() methods.  Stub out the handleAddFolder() and handleRemove() methods for later.  In render(), initialize the edit button dropdown toggle.  In addFiles(), create a DataONEObject for each file selected via the dialog or dropped in the dropzone.  Add each to the collection so they render.  Need to troubleshoot why they are defaulting to an EML formatId.