Support object format lookups
For brand new file uploads, set the object format id by searching the
object formats collection.  In getFormatId(), try to match the
`formatId` directly, the `mediaType`, or lastly the `extension` based
on the file name extension.  This needs work, but is a first pass.
Some extensions match multiple object formats, and currently it takes
the first in the list =|.
In createSysMeta(), use `parseHTML()` instead of `parseXML()` to be
consistent with other code, but I’m thinking we might want to revisit
using `parseXML()`.
Set the checksum and checksum algorithm properly.
also, add the `percentLoaded` attribute, which indicates how much of
the data file from the disk has been read.