Support reading files from disk
As a user selects files for upload from their disk, read each file into
memory (unfortunately). We have to generate a checksum for the file,
and I’ve yet to find a way to do so without fully reading the data into
memory.  We might be able to simulate an InputStream with
`FileReader.slice()` - needs work.
I’ve introduced a new `uploadStatus` using for ‘loading’ (l) since
reading large data files can take a while.  Testing shows it takes 10
seconds to read and md5 sum a 500MB file.  I may change this status ti
‘importing’ (i) to reduce confusion with file uploading to the server.
Once a file is read and checksummed, set the checksum and algorithm,
and change the status to queued in `handleLoadSuccess()`.  Add in other
reading event handlers (`updateLoadProgress()`, `handleLoadProgress()`,
`handleLoadError()`, `handleLoadAbort()`, and `handleCancel()`.  these
aren’t fleshed out, but are there for when we add progress bars and a
cancel button to the UI.