Determine content updates
DataONEObject objects instantiate with a limited set of attributes that
effectively only reflect the system metadata of an object, plus some
internal attributes to the object.  In initialize(), store these
attribute names as an array called `originalAttrs`. Register a listener
for all `change` events called `handleChange()`.  If an attribute not
in the `originalAttrs` is changed, we know the attribute came from an
object upgraded to `ScienceMetadata`, `EML211`, etc., which represents
a “content change”.  Set the `hasContentChanges` flag to true when we
have changes outside of the `originalAttrs`. On `save()`, reset this to
false.