Serialize system metadata fields in correct order
`serializeSysMeta()` is not producing XSD compliant documents, largely
due to ordering.  The `obsoletes` field is being appended to the end of
the document.  Reorder the DOM insertions, and add a bit more logic for
placing nodes.  Basically, the rule of thumb is to find the previous
sibling, and if it’s optional, find the next previous sibling, up until
we hit a required node.  I will continue this pattern for the rest of
the fields. Move the vars to the top of the function since they get
hoisted anyway. Oh, and always replace the submitter.