Fix bug in DataPackageView for new packages
Before this fix, navigating to #share would result in the Editor view
creating an EML211 and a DataPackage instance as expected but the
DataPackageView's table wasn't getting filled in. This was caused by
incorrectly adding the HTML for each row using $() instead of this.$el.
Using $() doesn't work because the HTML for the template isn't on the
DOM yet.

I also added an explicit call to addAll() so the models in the package
at init-time are rendered.