Change how the DataPackageView table gets filled in
This set of changes is intended to make it so the DataPackageView
renders properly for all three cases I'm working against right now,

1. Navigate to #share (no pid)
2. Navigate to #share on a metadata-only (no resource map) dataset
3. Navigate to #share on a normal Data Package (scimeta+resmap)

I'm not entirely sure this is the way to go but I've tested this change
and it works. Before this change, the scimeta row was being rendered
twice (two duplicated rows) when navigating to #share for an existing
package. This should fix it!