Fix a bug in the CitationView related to the title
In 7bdbe975e50a94b1d4e246c216da4726008252c2, support was added for the case where the metadata had multiple titles but it looks like a minor syntax error was introduced that resulted in the title not being displayed in the CitationView when the EML only has one title. This fixes that so rendering of the view should work in both cases.