Fix small bugs in setting parentModels on submodels
For child models being set with anonymous functions via _.each, the parentModel attribute was getting set incorrectly because of an incorrect use of this (which changes inside anonymous functions). I made a reference to the model outside the scope of the anonymous functions and passed that in as the parentModel when creating new child models.