Change how DataItem names (title/fileName) are edited
This changes the DOM for DataItem names from a td with contenteditable
set to true to input elements with placeholder attributes. The main
reason for this was because it's an easy fix for putting placeholder
text in which is harder to do on a td with contenteditable set to true.

The logic for when an update is triggered also changed a bit here. An
update will now only get triggered if the title is changed *and* that
change wasn't simply to clear it.