Support the full AccessPolicy schema
The DataONE AccessPolicy schema allows for each AccessRule to contain
one to many subjects with one to many permissions.  Our code was not
handling subjects as lists.  Refactor `serializeAccessPolicy` to
iterate through the stated subjects and the stated permissions, and
produce an allow rule for each pair.  This ensures we handle every
rule, even though it repeats subject or permissions.