Support arrays or objects for AccessRules
When there is a single AccessRule in the AccessPolicy, `_.each()` is
passed an Object, but when there are more, it’s passed an Array.  Be
sure we always have an array no matter home many rules are in the
policy so we can access the subject and permissions attributes
consistently.