Support adding files to new packages ...
When we start a package from scratch, we already create the DataPackage
instance and the EML211 instance, but we don’t link them together.
This causes `DataItemView.getParentDataPackage()` to return an empty
array, so we can’t add new objects to the package.  So, on
instantiating the package, link it to the science metadata with the
`resourceMap` attribute.