Support round trip updates on fresh packages
When we create a fresh package, isNew() returns true, so we POST both
the EML and RDF to the server.  However, once this is complete, the
isNew() was still returning true on a subsequent edit because we don’t
have the dateUploaded value that is set by the MN.  So, on success,
call fetch() for each objects’ system metadata and merge the values
into the model.

Likewise, we still need the handleChange() callback to fire when
attributes are changed so we set the uploadStatus to “q” which
determines if changes will be save()’d to the server.  Lauren, I see
that we still have major performance issues with this, so I’ll work on
tracking it down too.  Your and Bryce’s review of this would be helpful.

As an aside - for EML and RDF, we weren’t setting the fileName field
correctly, so add DataONEObject.setMissingFileName() which gets called
prior to serializeSysMeta() so we always have some reasonable file name
(defaults to pid + “.” + extension from the ObjectFormats collection).