Serialize textModels correctly
It seems we have a problem with the JQuery replaceWith() call, and so
the abstract node was not being replaced with the contents from
updateDOM().  With Lauren and Bryce’s help, we found a workaround using
.html().  This may be problematic when needing to also transfer node
attributes like the id in EMLParty, so I’ll be tracking down the
replaceWith() issue, which may be related to the eml document DOM not
being part of the HTML document DOM in memory.