Allow EML211 object to fetch sysmeta
When we save an EML211 object to the MN, it is still considered new
until the system metadata has been fetched and the dateUploaded field
is populated and not null.  Don’t check for isNew() in
fetchSystemMetadata() because it will always return true on the first
save. Subsequent saves will return false for isNew().