In updateID(), also update other attributes
When we change are updating one object to the next and call
DataONEObject.updateID(), we set the oldPid and generate a new one. For
EML211 documents, we rely on the documents and isDocumentedBy
attributes to keep track of those relationships when building the
package. Update these fields as well when we update the id so we
serialize the DataPackage and don’t miss removing the old statements
and replacing them with new ones.