Add support for the dirty metadata flag
Add a `toggleControls()` method that will toggle the Save controls
based on the status of the package model’s `changed` attribute.  When
`saveSuccess()` is called, toggle the controls and make sure listeners
are set.  In setListeners(), now have the EditorView listen to changes
on the package model’s `changed` attribute.

refs https://github.com/NCEAS/metacatui/issues/44