Add support for a dirty metadata flag
Add the `changed` attribute to indicate a change in the sysmeta,
scimitar, or data package to drive the toggling of the Save bar in the
editor. Fix `handleChange()` so it discerns between content, system
metadata, and non-content-or-system-metadata attribute changes. Also,
reorder attributes just to organize them as sysmeta attributes or not.
Also, add the objectXML attribute to DataONEObject because it gets
`set()`  during the object upgrade to EML, and we want to ignore that
in terms of the dirty flag.

refs https://github.com/NCEAS/metacatui/issues/44