Fix the contenteditable filename field not being editable on IE11
The solution here was to figure out that IE11 can't do contenteditable
on <td> elements but it can do it on divs. So I throw a div inside the
filename <td> that's contenteditable which appears to work across
browsers. Tested with Chrome/Firefox/Safari on Mac, Edge and IE11 on
Windows 10.

Closes #52