Fix serialization issues
Set the latestVersion attribute when initializing the package so it’s
always available. When finding the metadata in a package, look through
this.models rather than this.get(“members”). To deal with repeated
triple statements written to the RDF, add conditional statements to
only add the triple if it doesn’t already exist.  Change some minor
formatting.