Deal with multiple aggregations in RDF
The DataPackage.serialize() has been creating incorrect RDF by not
removing the old aggregation during a package update, and so when a
package is refreshed from the server, an empty aggregation is rendered,
when the second aggregation in the RDF actually aggregates the members
(not the first).  The following changes improve this:

- Use cloned lists of statements when iterating over RDF statements and
modifying/removing them so the list doesn’t change mid stream.
- Remove statements from the RDF first, modify the locally cloned
statement, then add it back in. In situ modifications seem to fail.
- Use try/catch when removing to avoid errors when statements have
already been moved, like in removeFromAggregation().

Please pardon the whitespace changes throughout the file - I switched
to using Atom and apparently it does that - need to tweak settings.

Also, the calls to statementsMatching() in addToAggregation() seem to
not be working, and so we are still adding duplicate triples into the
RDF. I’ll be looking tracking that down, but it seems to have to do
with something internal to rdflib.js.

refs https://github.com/NCEAS/metacatui/issues/66