Fix more statement removals
Failures in calls to statementsMatching() inside of
IndexedFormula.remove() seem to be associated with destructive changes
in the underlying array of graph statements. When removing
isAggregatedByStatements, also clone the resulting list so as to not
mess with the underlying structure.   This looks like the last instance
of this happening - all other lists are cloned before iterating and
removing.