Remove repeating statements from the RDF
This is tricky because of what seems like references to Statement
objects get modified across the IndexedFormula index because all
references within the IndexedFormula point to a single Statement
object.  When a call to IndexedFormula.remove(statement) doesn’t find
the statement in the index because the subject, predicate, or object
node in the statement  was previously altered in another call, the
desired statement to be removed doesn’t get removed, and then we
subsequently add another statement in with the new pid replaced in the
subject or object value.  To mitigate this, add in the DataPackage.rdf
attribute so we have access to rdflyib from this.rdf, and in
serialize(), clone nodes in statements with
this.rdf.Node.fromValue(statement.{subject|predicate|object}) so we can
remove statements by value rather than by reference.  Note that
Jquery.extend() only clones simple objects and arrays well, so I didn’t
use it in this case. This seems to work for all statements except the
ORE:describes statements.  Still tracking that down, and I may just
remove the statement before any modifications of the graph since it can
be added in at any point with the new pid value.