Prevent repeat describes/isDescribedBy statement
To mitigate this, remove all ore:describes and ore:isDescribedBy
statements at the beginning of DataPackage.serialize(), and then add
them back in with the new pid for the package.

Also, use `this.rdf` consistently rather than just `rdf` since we need
access to it via the `this` context passed into calls to `_.each()` etc.