Write the EMLEntity.spec tests
To get started with the testing framework, create a spec file for the
EMLEntity object.  So far, est the constructor and the parse() method.
To seed the response for parse(), I created the EntityUtil object and
added a getTestEntityXML() method.  We’ll need to think about how we
want to do this - perhaps use a shared JS file across tests - we’ll see.