Create an EMLMeasurementScale factory
EMLMeasurementScale provides a static factory method for producing an
appropriate measurement scale object based on the incoming
<measurementScale> XML fragment.  For nominal and ordinal scales, it
returns an EMLNonNumericDomain object.  For interval and ratio scales
it returns an EMLNumericDomain object.  For a dateTime scale, it
returns an EMLDateTimeDomain object. This factory is used in
EMLAttribute.parse() to create the correct measurementScale attribute.

refs https://github.com/NCEAS/metacatui/issues/87