Improve EMLEntity.parse()
The parse() method now accepts an attributes object (like other
Backbone models), but expects the XML string to be set on the
attributes.objectDOM property.  This allows us to save the objectDOM in
the resulting attributes that get instantiated into the object. Fix the
unit test to pass in an object.

Also, add addAttribute() and removeAttribute() methods for managing the
attributeList for any entity.  Tests pending.

Lastly, make trickleUpChange() a class method so it can be called
statically as a callback.

refs https://github.com/NCEAS/metacatui/issues/87