More work on EMLEntity:
Add the nodeOrder array attribute, along with the getEMLPosition()
method that can be used by this and subclasses to find the position to
append after.  Use attributes.objectXML in parse(), and default to
appending attributes in addAttribute() when the index argument isn’t
passed in. Add formatXML() to for use with this and subclasses.

Implement updateDOM() for the basic fields. More to come as we fill in
the model with sub sections like EMLPhysical, etc.

refs https://github.com/NCEAS/metacatui/issues/87