Update unit tests
Use attributes.objectXML when creating instances and passing in the
attributes.  For EMLEntity, note that there are variable scoping issues
WRT Mocha syntax.  It looks like variables are available within each
describe() block, but not in nested describe() blocks, so I’ve moved
variables to the inner describe() calls - a little more repeated, but
it works fine.