Implement EMLNumericDomain.parse()
Parse the measurementScale XML fragment being passed into
EMLNumericDomain, which gives us units, precision, numericDomain,
numberType, bounds, minimum, and maximum. Note that the unit attribute
is returned as an object with either a standardUnit or customUnit
attribute. Also, for the single numericDomain attribute, it is an
object with a required numberType attribute, and an optional bounds
attribute, which is an array of objects with minimum and/or maximum
attributes.

refs https://github.com/NCEAS/metacatui/issues/87