Create an EMLUnit object
An EMLUnit represents a unit listed in the EML Unit Dictionary.  By
parsing the XML with X2JS, attributes of a unit are converted to JSON
fields with an underscore prefixed to them to enable re-serializing
back to XML, so I’ve kept them that way for this simple schema.

refs https://github.com/NCEAS/metacatui/issues/113