Fix EMLMeasurementScale unit tests
EMLMeasurementScale no longer needs to use a DOMParser since we have a
temporary hack for <source> elements. When creating new instances, pass
in the objectDOM, not objectXML.

Note: @laurenwalker I see you changed the factory to produce objects by
passing in the domainName only, with no XML fragment. I’m thinking that
the factory should be used for existing content where we pass in the
fragment, and that we directly instantiate EMLNonNumericDomain,
EMLNumericDomain, and EMLDateTime using their respective constructors
directly if we a starting with fresh new objects.  Let’s discuss.