Implement updateDOM() for textDomain content
Update the objectDOM with values from the Backbone object attributes.
Determine if the domain is textDomain or enumeratedDomain, and if it
pre-exists with an id or if it is new content.  Either update existing
content or append new content as needed.  So far, this is working for
textDomain - I’ll be adding in enumeratedDomain next.