Add attributes to the DOM in the EMLEntity superclass
First, if the entity has no id attribute, generate one so we can keep
track of entities by id. Change insertAfter to nodeToInsertAfter for
clarity.  Given the attributeList array in the entity, get the
attribute ids, and check to see if they exist in the DOM.  If so,
update (replace) them by id (TODO), otherwise, when they don’t exist,
add each new attribute to the DOM.