Add in the unit information in updateDOM()
Units are represented by an object like {standardUnit: “celsius”}, so
parse the object to get the standard or custom unit and add it to the
DOM.  When we support custom units, we’ll need to update the EML
additionalMetadata with the unit element and refer to it by id.

Also, domains are <xs:restriction> on a <SimpleType>, so can’t have an
id
attribute. Remove it.