Update entity attribute list by position not identifier
After discussing this with Lauren, we’ve decided to not assign
identifiers to each attribute and update-by-id, but rather clear the
DOM attribute list and iterate through the EMLEntity.attributeList
backbone models and recreate the list in the DOM.  We’re able to do
this because each EMLAttribute has cached its objectDOM, which we
update without destroying. Across saves, the information in the
objectDOM, combined with any updates in the Backbone model, should
provide a complete view.  When the page is refreshed, the objectDOM is
cached again with the newest edits.