Replace the measurementScale subtree
During EML serialization, for each EMLAttribute that we call
updateDOM(), we repopulate the measurementScale.  Don’t just append it
here, but remove the old one and then append it(), otherwise we double
up on measurementScale domain elements.

Also, when setting the objectDOM attribute, set the DOM itself, not the
jQuery object, which was causing serialization issues.