Handle entities correctly in getEMLPosition()
EML211.getEMLPosition() has been returning the wrong position for
adding new entities like `otherEntity`, `dataTable`, etc.  The
nodeOrder array lists these entity types in the order they are found in
the EML schema, but in fact they can be appended (optionally) in any
order and can repeat ( {0..n}+ ). To pass back the correct node to
insert after, getEMLPosition() now looks to see if the given nodeName
is one of the entity type names, and if so, returns the last child of
the <dataset> node. So, it effectively appends the node rather than
inserts the node.

refs https://github.com/NCEAS/metacatui/issues/169