Add the EMLDataTable class
The EMLDataTable models the eml-datatable module and extends the
EMLEntity class, basically just adding in the caseSensitive and
numberOfRecords fields.  It also requires the attirbuteList.  For now
in validate() we check that the list is non-zero, but we may need a
more comprehensive check (that the number of attributes are correct
with regard to the data file being described).  This may require the
MDQ engine.

refs #180