Fix minor bug in EML211.parse() for entitites
EMLEntity is really an abstract superclass for all other entities
within EML.  If we instantiate it, we will end up serializing invalid
EML documents.  So, as a fallback when we don’t know what the entity
type is based on the localName of the element, instantiate an
EMLOtherEntity with an entityType of “application/octet-stream”, which
will at least produce valid EML.  To really fix this, we need to
support all entity types listed in #161.