Changed GeoLocation Error Wording
Part of #254 includes changing the wording of the lat/long validation warnings.

"Refer to values as latitude and longitude coordinates, not North,West,South,East coordinates
We are asking users to enter a bounding box by supplying the coordinates of the Northwest corner of that box and the Southeast corner. So if the Northwest latitude value is wrong, the error message should word it as the "Northwest latitude", not "North bounding coordinate." I think this will keep wording consistent and avoid confusion."