Change Validation Procedure
This is for issue #254   
Validating latitude/longitude boxes between clicks happens in updateModel (EMLGeoCoverageView.js). 

Validating lat/long when the user clicks outside an element occurs in showValidation (in *View.js). I pulled some code out of showValidation because the same code was needed in updateModel. I created showValidationError, and removeValidationError from this code.

The code referenced above makes calls into EMLGeoCoverage.js to handle the actual validation. 

validate: function() is called from showValidation when the user clicks outside. This was originally only displaying one error at a time, which I extended to show all of them. This was done by creating errMsg and instead of immediately returning when we hit an error, keep going and append the errors to the string.

There was some code inside validate() that was needed by updateModel, so I modularized some of the code into getCoordinateStatus, generateStatusErrors, and getErrorMessage.