Location Input With Persisting Borders
This issue adds borders that stay on when the user clicks outside the lat/long boxes. It also keeps the error messages from disapearing and shows all error messages instead of one.

EMLGeoCoverage.js changes

Extended addspace: function to append to the front of the string. Changed the check on the string length because we need to support the case when ‘msg’ is undefined.

Added the rest of the error messages to getErrorMessage: function to allow access from the view and other places in EMLGeoCoverage.js

I needed to check for coordinate pairs in two places, so I took the subroutines out of validate: function and turned them in to checkForMissing: function and checkForPairs: function. To keep the new methods general and true to their names, I took out the error message references and put that logic on the caller.

EMLGeoCoverageView.js Changes

Changes to updateModel: function
	Split a few if statements
	Moved the validation code out into the function validateCoordinates

The validateCoordinates function handles the case where the user is clicking between latitude and longitude boxes. It includes everything from error message display to border highlighting.

Changes to validateRow: function
At this point we can just use the `status` object to check for validity. I created a new method that adds the error class to the correct (invalid) input box. In validateRow, this was the long else if chain. The code was put in getBorderStatus, and the ‘else’ statements were removed because we want to display all of the error messages. validateRow calls addErrorClasses, which uses the result of getBorderStatus to add the classes.

This was added as part of issue #254