Make Metadata Tile Editable After Save
This is a fix for issue #349 where the filenames and title were not editable.

The problem was that we were not setting the title to editable after saving the data. We also implimented the rename functionality so I removed the event and function.

On linw 594 we had already defined output so I removed the `var`.

The fix occured on lines 672-675