Use a cross-browser DOMTokenList API methods
In `ProvChartView.selectProvEntities()` the call to
`e.currentTarget.classList.values()` fails under IE 11 because the
DOMTokenList API implementation doesn't include that method.
All browsers seem to implement the `e.currentTarget.classList`
property as a DOMTokenList though, and since it is an array, we can
call `contains()` on all platforms.  For some reason, assigning the list
to the `myClasses` variable results in `undefined` in IE, so we use
the full `e.currentTarget.classList.contains()` instead.

refs #376