Locally serve the AnnotateIt dependencies
The annotator.js (full and min) libraries make a call to $.getScript()
over HTTP not HTTPS, and so IE flags the insecure content. Mitigate this by:
adding the dependencies to the components directory and updating the 1.2.10
annotator.js files to pull from the local location instead.

Note that this may be a breaking change.  I don't get an error on
MetacatUI initialization, but this needs testing with the annotator
feature enabled.  The bigger problem is that AnnotateIt has closed
shop, so we need to migrate away from it anyway.

Also, remove the 1.2.9 version - we don't depend on it now.

refs #376