Add an app-level event bus
In app.js, add the MetacatUI.eventDispatcher attribute that is a
BackBone.Events object and acts as an event bus across components of the
application. This can be useful to organize our events (if we coordinate
on that), but in this use case, I need to respond to events on a model
that isn't added to the collection, and so can't be referenced.
Triggering named events on the eventDispatcher and responding to them
via callback is the solution.

refs #382