Dereference references
Handling references is not simple in that not only do we need to be able
to keep the DOM up to date when a referenced node is deleted or altered
in some way, but we also need to support references in many of our
views, which we don't yet.

To get editing of existing content with <reference> elements working in
the short term, I've added an `EML211.dereference()` method that is
called at the beginning of `EML211.parse()`. This method dereferences
the nodes and copies the referenced content into the referencing
location.

This is really a stop-gap measure, but I will add another ticket to
properly handle references in all of our views and models.

refs #295