Fix one more package regression
I introduced a bug in in #292. To fix it, call handleAdd() when the
successLoading event is fired. We can't do this in the DataPackage
itself because when existing packages get populated from the server, we
don't want it to modify the EML or RDF. This can only happen at
DataONEObject.create() time.