Fixed bug where EMLEntities were matched with more than one DataONEObject, creating duplicate entity XML node ids during serialization.
This was done by:
Save a reference to the DataONEObject model in the EMLEntity model when
we successfully matched a DataONEObject to it's EML Entity.

In EML211.getEntity(), when we find a matching entity (via id, checksum,
file name, etc.), make sure it hasn't already been paired with another
DataONEObject model. If it has, do not match it again.

When a new file is added to the package, do not create an new EMLEntity
if there already is one for that DataONEObject.