Fix remaining issues with app resources not loading
We noticed that top-level routes, like `/data` worked, but that routes at lower levels, like `/view/{PID}` did not. This was because the request was trying a path that doesn't exist. This makes sense. Just switching the Require.js baseUrl from `js/` to `/js` did not entirely fix it so I had to change paths in all of the places I've done in this commit.

To test that the changes in this commit were the minimum necessary changes, I reverted all of my changes, then made each change one-at-a-time.