Add copy-on-click support to the DOM returned by the view service (#592)
This behavior is slightly different than the Copy Citation in that it just replaces the innerHTML with a ✔ and, after 500ms, returns the old button (or other element) innerHTML. This'll also work if someone needs copiables in other areas of the view service later on.

Note from Lauren during PR review: I am going to move the call to insertCopiables() into the renderMetadata() function to follow the design of the view, which is to only use render() to set up the basic components of the view and start the fetch the model. All of the detailed elements that are added to the metadata rendering is done in renderMetadata().