Fix bug in MetadataView.insertCopiables
I had some bad logic in my initial implementation that caused the button text to permanently become a ✔️ if the user clicks the button fast. This was because the callback that re-sets the text back to its original value picks up the termporary value. Now the button will always return to "Copy".