Add support for Highwire Press bibliographic meta tags
This change required a refactor of the JSONLD work to DRY up code that needed similar logic between the JSONLD and Highwire Press code.

While I was in here, I realized we weren't properly removing stale JSONLD so I added a method that is attached to all routers' 'route' event that clears the stale JSONLD and also the Highwire Press meta tags.