Nope. Had to move collection.fetch back inside ProjectModel.fetch.success
I'd confused the id ProjectModel ID and the CollectionModel ID. I was trying to use the ProjModel one to fetch the collectionModel. That didn't work (obviously - I just got two copies of the Project XML). In order to use the correct CollectionID I have to wait for success on the projectModel fetch so that I have the collectionID. But, by waiting, it means I don't have access to that collection json fetch by the time I'm in the project model parse function. Out of luck both ways. #798