Initial commit of ProjectHomeView and template.
This is in reference to #63 and #35. I lifted the template from the previous version of the project.html template. In doing so, I renamed `project-main` to `project-home` to be more consistent with what we've been calling that tab. I've also made the assumption that the `project-tabs` container div is going to be in another template, but I'm not sure exactly where it should go.