Begin the project pages feature technical documentation
The project pages will be implemented as a group of subviews of the 
ProjectView, which is rendered from the AppView in the #Content 
container.  Add UML diagrams and documentation that explains our design 
decisions to date, and add in the wireframes and mockups.  The 
ProjectHomeView documentation is in, and I'll continue to add other 
subviews.

refs https://github.com/NCEAS/project-papers/issues/35