Progress on roughing out the project views and nested project section views.
This was looking a bit closer to the wireframes before Lauren's changes to the css, but it still serves. There's currently just a single tab representing a project section. ...and it's an instance of project section. The next step is to replace that instance with project home, people, metrics, and visualizations that sub class project section view. NCEAS/project-papers#35 #824 #815