Some more progress on nesting views.
I'm having some confusion about getting the tabs to work properly. I'm not sure if it's to do with my sketchy understanding of bootstrap or if it's to do with the new routing. I'll work on it more soon. I made a very minor addition to the css, but I'll probably reverse that once I figure out how to use bootstrap properly for the layout. NCEAS/project-pages#35 #829