Got markdown from model to render in markdownView; simplified projectView
This is the first step in setting up the markdown view. The next step is using showdown to convert the markdown to HTML. I added a listener to projectView so that the subviews render after the model syncs, since information from the model is required for all the subviews. This causes a delay before the subviews are rendered, so perhaps there is a better way to achieve this. I added all the subviews to the subviews array and used _.invoke to have them all render.

refs https://github.com/NCEAS/metacatui/issues/815, https://github.com/NCEAS/metacatui/issues/820