Enable routing to named projects
Change the UIRouter to match the following path:

projects(/:projectId)(/:projectSection)

where projectId is the optional pid, seriesId, or projectName (like 
"SASAP").  Change UIRouter.renderProject() to look up projectIds based 
on given projectNames, or vice versa.  When a projectSection is given, 
route to it and also pass it to the ProjectView for rendering.

Note the change from "project" to "projects" in the path (based on the 
ticket discussion).

This will be a breaking change until I update ProjectView to handle 
these parameters.  The previous implementation was saving project state 
in MetacatUI.appView.projectView, which I think we need to avoid.

refs https://github.com/NCEAS/project-papers/issues/65