ProjectMembersView now working.
Things got a little awkward trying to create the fluid row divs, but I'm hoping the way I've done it is acceptable. I've provided a spot for phone number in the template, but none of the associated parties in the example project have phone numbers (at least they're not in the model output) so phone number display is untested at this point. #825