Add StatsView in metrics tab of project page
Still need to correctly set the search model. In the SASAP example at the moment, the stats view shows fewer datasets than in the dataCatalogView, and 0 datafiles. The metrics tab uses the StatsView which is called directly from the project view. Even though I did not use the ProjectMetricsView or ProjectMetricsTemplate, I left those files here in case we want to create the stats view in the Metrics view instead.

relates to: https://github.com/NCEAS/metacatui/issues/823