Fix bug where line chart in project metrics view was squished
This problem was related to the timing of rendering the metrics view. The line chart in stats view inherits it's width from the parent div in the template (class: span6). The parent div seems not to have it's full width until the user navigates to the metrics tab (or is delayed for some other reason). I fixed this by only rendering the metrics view once the user has clicked on the metrics tab. This event is only listened to once. This seems to solve the problem.

relates to https://github.com/NCEAS/metacatui/issues/823#issuecomment-448715113