Added project logos to bottom of ProjectMembersView.
Since different projects seem likely to have different numbers of logos, I made it so that we can display between 1 and 4 logos on a single row with appropiate spacing (using `row-fluid`). With more than 4 logos, we'll have multiple rows of 4. I've added the logos to the ProjectMembersView (#825). It should be pretty trivial to add them to the bottom of the ProjectHomeView as well (#824 and #730).