Truncate second level items in TOC.
If second level items in table of contents (derived from H2 tags) are more than 15 characters long, they will be truncated first whitespace after the 15th character and `...` will be added. ...except when the truncated for (with `...` added) is longer than the original item. That will happen when the 15th character is within 3 characters of the end. #818