Override the render method
Add in a render method that is specific to a data catalog view that uses 
the filters collection. 
Also, revert back to using "searchResults" rather than "collection" for 
the name of the main collection of the view to stay consistent with 
DataCatalogView (this diverges a bit from the Backbone semantics, but is 
not a big deal).  
Start to add searchModel references back in.  I've been debating if we 
should nix the use of "models/Search" in this view -- in favor of 
storing all filters in Filters, and all other attributes in this view 
itself, but I now think they can play well together.  Meaning, the 
Filters collection can augment the Search model, rather than replace it 
completely.

refs #846