Incorporate the Search model into the ProjectModel
After putting the Filters collection as a property of the Search model, 
we need access to it in the ProjectView (and its subviews).  I've added 
it into the ProjectModel, and when we parse the filterGroups from the 
project document, we add each filter into the Filters collection, but 
now augment the searchModel with a filters attribute instead of setting 
it directly in the project model.

Also, cache the project model in a MetacatUI.projects hashmap by project 
id rather than an individual project since the UI will have multiple 
projects associated with it, and we can cache the state of each as 
needed.
refs #846